test(e2e): 三份合成一个窗口,启动次数 5 → 3
smoke / schedules / archive 要的环境完全一样(已登录 + 干净 HOME),却各起一次
Electron。实测每次 launch+close 约 3.2s(launch 1.3s、firstWindow 1.4s、
等界面就绪 0.3s、close 0.2s) —— 纯开销占了总时长约一半。合成 workbench.mjs,
按「外壳 → 导航 → 弹窗 → 定时任务 → 会话」顺序跑完,段与段之间自己收拾干净
(菜单 z-index 40,留着会挡住下一段要点的按钮)。

另两份**不能**合,合了就验不到原本要验的东西:
- login.mjs 刻意不设 WYCKOFF_E2E_FAKE_SIGNIN,要的就是未登录的启动路径
- artifact-host.mjs 要在主进程自建 WebContentsView 和本地 STUN 服务器

断言一条没少:合并前后逐条比对过名字,50 条对 50 条。中间真丢过一条
(「设置弹窗打开了」被 openSettings helper 吞掉了),是比对发现的,已补回。

## 顺带修掉一个改动前就存在的 flake

login.mjs 的「焦点进入弹窗」本机 5 轮红 4 轮:`waitFor visible` 只保证元素可见,
而自动聚焦是 React 挂载后的 effect、晚一帧才跑,紧接着 evaluate 读到的
activeElement 常常还是 body。改成 waitForFunction 轮询到位再断言,6/6 绿。
失败文案也带上实际的 activeElement —— 原来只报正则不匹配,看不出是没聚焦
还是聚焦错了元素。

## 顺带发现 schedules 和 archive 根本没进 CI

desktop.yml 里只跑了 smoke 和 artifact-host。合并之后它们跟着 workbench 一起
跑上了 —— 定时任务和会话归档这两条链路以前只在本机验过。

## 实测

启动次数 5 → 3。墙钟 37-40s → 26-36s,方差不小(本机噪声),别把它当成稳定
的百分比。真正确定的是少了两次进程启停和两份重复的 launch/check/种数据代码。
